chore: Refactor types to maintain code modularity
