+++ src/student_registry_project/implementations/course_registry.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use crate::student_registry_project::{
-    types::basic_types::{Course, CourseRegistry},
+    types::{course::Course, course_registry::CourseRegistry},
     utils::convert_to_string,
 };
