+++ src/student_registry_project/types/student_registry.rs
@@ -1,0 +1,7 @@
+use super::{course_registry::CourseRegistry, student::Student};
+
+#[derive(Debug, Clone)]
+pub struct StudentRegistry {
+    pub total_students: Vec<Student>,
+    pub course_registry: CourseRegistry,
+}
